backend/requirements.txt
asgiref==3.9.1
async-timeout==5.0.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
django-ckeditor==6.7.3
django-cors-headers==4.7.0
django-filter==25.1
django-js-asset==3.1.2
django-redis==6.0.0
django-tinymce==4.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
Markdown==3.8.2
pillow==11.3.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
redis==6.4.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
WooCommerce==3.0.0
